Move oonireport, ooniresources and oonideckgen into the scripts/ directory.
* Delete unused oonid command

* Create ooniprobe-agent script
