+++ ooni/scripts/oonireport.py
@@ -1,16 +1,20 @@
 from __future__ import print_function
-import yaml
-import sys
 
-from twisted.internet import defer
+import os
+import sys
+import yaml
+
+from twisted.python import usage
+from twisted.internet import defer, task
 
 from ooni.constants import CANONICAL_BOUNCER_ONION
 from ooni.reporter import OONIBReporter, OONIBReportLog
 
 from ooni.utils import log
-from ooni.report import parser
 from ooni.settings import config
 from ooni.backend_client import BouncerClient, CollectorClient
+
+__version__ = "0.1.0"
 
 @defer.inlineCallbacks
 def lookup_collector_client(report_header, bouncer):
@@ -41,7 +45,7 @@ def upload(report_file, collector=None, bouncer=None):
     with open(config.report_log_file) as f:
         report_log = yaml.safe_load(f)
 
-    report = parser.ReportLoader(report_file)
+    report = ReportLoader(report_file)
     if bouncer and collector_client is None:
         collector_client = yield lookup_collector_client(report.header,
                                                          bouncer)
@@ -115,3 +119,120 @@ def status():
     print("------------------")
     for report_file, value in oonib_report_log.reports_incomplete:
         print_report(report_file, value)
+
+class ReportLoader(object):
+    _header_keys = (
+        'probe_asn',
+        'probe_cc',
+        'probe_ip',
+        'start_time',
+        'test_name',
+        'test_version',
+        'options',
+        'input_hashes',
+        'software_name',
+        'software_version'
+    )
+
+    def __init__(self, report_filename):
+        self._fp = open(report_filename)
+        self._yfp = yaml.safe_load_all(self._fp)
+
+        self.header = self._yfp.next()
+
+    def __iter__(self):
+        return self
+
+    def next(self):
+        try:
+            return self._yfp.next()
+        except StopIteration:
+            self.close()
+            raise StopIteration
+
+    def close(self):
+        self._fp.close()
+
+class Options(usage.Options):
+
+    synopsis = """%s [options] upload | status
+""" % (os.path.basename(sys.argv[0]),)
+
+    optFlags = [
+        ["default-collector", "d", "Upload the reports to the default "
+                                   "collector that is looked up with the "
+                                   "canonical bouncer."]
+    ]
+
+    optParameters = [
+        ["configfile", "f", None,
+         "Specify the configuration file to use."],
+        ["collector", "c", None,
+         "Specify the collector to upload the result to."],
+        ["bouncer", "b", None,
+         "Specify the bouncer to query for a collector."]
+    ]
+
+    def opt_version(self):
+        print("oonireport version: %s" % __version__)
+        sys.exit(0)
+
+    def parseArgs(self, *args):
+        if len(args) == 0:
+            raise usage.UsageError(
+                "Must specify at least one command"
+            )
+            return
+        self['command'] = args[0]
+        if self['command'] not in ("upload", "status"):
+            raise usage.UsageError(
+                "Must specify either command upload or status"
+            )
+        if self['command'] == "upload":
+            try:
+                self['report_file'] = args[1]
+            except IndexError:
+                self['report_file'] = None
+
+
+def tor_check():
+    if not config.tor.socks_port:
+        print("Currently oonireport requires that you start Tor yourself "
+              "and set the socks_port inside of ooniprobe.conf")
+        sys.exit(1)
+
+
+def oonireport(args=sys.argv[1:]):
+    options = Options()
+    try:
+        options.parseOptions(args)
+    except Exception as exc:
+        print("Error: %s" % exc)
+        print(options)
+        sys.exit(2)
+    config.global_options = dict(options)
+    config.set_paths()
+    config.read_config_file()
+
+    if options['default-collector']:
+        options['bouncer'] = CANONICAL_BOUNCER_ONION
+
+    if options['command'] == "upload" and options['report_file']:
+        tor_check()
+        return upload(options['report_file'],
+                      options['collector'],
+                      options['bouncer'])
+    elif options['command'] == "upload":
+        tor_check()
+        return upload_all(options['collector'],
+                          options['bouncer'])
+    elif options['command'] == "status":
+        return status()
+    else:
+        print(options)
+
+def run():
+    task.react(oonireport)
+
+if __name__ == "__main__":
+    run()
